refactor: LottoService 네이밍 변경
